updater/requirements.txt
crc
pyserial